requirements.txt
beautifulsoup4==4.12.3
biopython==1.83
biotite==0.40.0
bs4==0.0.2
fair-esm==2.0.0
idna==3.4
igfold==0.4.0
lmdb==1.6.2
logomaker==0.8
ml-collections==0.1.1
scikit-learn==1.4.2
seaborn==0.13.2
six==1.16.0
tqdm==4.66.2
torch==2.6.0